requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==1.11.3
django-celery==3.2.2
django-crontab==0.7.1
django-simple-captcha==0.4.6
idna==2.8
kombu==3.0.37
mysqlclient==1.3.13
Naked==0.1.31
Pillow==5.3.0
pycparser==2.19
pycryptodome==3.7.2
PyMySQL==0.9.2
python-alipay-sdk==1.8.0
pytz==2018.7
PyYAML==3.13
redis==3.0.1
requests==2.21.0
shellescape==3.4.1
six==1.11.0
urllib3==1.24.1
